Fix dropdown resetting under layer-shell
window_setup unconditionally recreated the entire input section every
time it was called. As window focus results in window_setup being called
when layer-shell is active, this mean that the focus change caused by
interacting with the dropdown would result in the dropdown being
destroyed and recreated, resulting in it being reset to the initial
value.

Only recreate the input section if question has changed to avoid resets,
and copy the active dropdown selection index if output focus changes to
improve usability.

Live edits to the dropdown entry are still lost on output change.